llama2/requirements.txt
langchain
pypdf
unstructured
sentence_transformers
pinecone-client
llama-cpp-python
huggingface_hub
openai
flask
